LayoutTests/imported/w3c/web-platform-tests/css/css-text/text-spacing/tentative/parsing/text-spacing-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: parsing text-autospace with valid values</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value('text-spacing', 'auto');
test_computed_value('text-spacing', 'none');
test_computed_value('text-spacing', 'no-autospace auto');
test_computed_value('text-spacing', 'auto no-autospace', 'no-autospace auto');
test_computed_value('text-spacing', 'no-autospace space-all', 'none');
test_computed_value('text-spacing', 'space-all no-autospace', 'none');
test_computed_value('text-spacing', 'auto space-all');
test_computed_value('text-spacing', 'space-all auto', 'auto space-all');
</script>
</body>
</html>
